# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Avances",
        "description": "Documentação gerada com L5-Swagger e OpenAPI 3.0",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/inventory-movements": {
            "get": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Listar movimentos de estoque",
                "description": "Retorna todos os movimentos de estoque filtrados por produto, data, etc.",
                "operationId": "fbf32c9cd1154f25ec3c96b97f2e7ff5",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limite por página",
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca pelo nome do produto",
                        "schema": {
                            "type": "string",
                            "example": "Parafuso"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filtrar por ID do produto",
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "example": "2025-10-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "example": "2025-10-20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de movimentos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InventoryMovement"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao listar movimentos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Registrar novo movimento de estoque",
                "operationId": "20bb9d5476177b2ee3a611c1a2078e58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryMovement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movimento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryMovement"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno ao criar movimento"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-movements/{id}": {
            "get": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Exibir um movimento específico",
                "operationId": "85bbcaa55ee44009c3153c96b58e0e39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do movimento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventoryMovement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movimento não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Atualizar movimento de estoque",
                "operationId": "c58504bb2389028df03486c341215c8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InventoryMovement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Excluir movimento de estoque",
                "operationId": "16b6d468f6289fc680677e4b7afd7aee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimento excluído com sucesso"
                    },
                    "404": {
                        "description": "Movimento não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-movements/{id}/restore": {
            "patch": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Restaurar movimento excluído (soft delete)",
                "operationId": "1295c3c55974f2058942ab58e8869ef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurado com sucesso"
                    },
                    "404": {
                        "description": "Movimento não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-movements/product/{productId}": {
            "get": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Listar movimentos por produto",
                "operationId": "f2dba4ac7258f42d2fdcdea87d3647b8",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimentos retornados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryMovement"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-movements/warehouse/{warehouseId}": {
            "get": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Listar movimentos por armazém",
                "operationId": "d4ea8efcec821047856b94eae0b02f3e",
                "parameters": [
                    {
                        "name": "warehouseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movimentos retornados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventoryMovement"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory-movements/stock": {
            "post": {
                "tags": [
                    "Inventory Movements"
                ],
                "summary": "Obter saldo atual por produto e armazém",
                "operationId": "38f8335c6e2ee0284ac4257d339089ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "warehouse_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saldo retornado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "warehouse_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "current_stock": {
                                                    "type": "number",
                                                    "example": 150.25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InventoryMovement": {
                "title": "Inventory Movement",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "warehouse_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "movement_type": {
                        "description": "ID do tipo de movimento",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity_movement": {
                        "type": "number",
                        "example": 10.5
                    },
                    "quantity_total": {
                        "type": "number",
                        "example": 100
                    },
                    "notes": {
                        "type": "string",
                        "example": "Entrada de produto"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:30:00Z"
                    },
                    "product": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 5
                            },
                            "name": {
                                "type": "string",
                                "example": "Produto Exemplo"
                            }
                        },
                        "type": "object"
                    },
                    "warehouse": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "name": {
                                "type": "string",
                                "example": "Depósito Central"
                            }
                        },
                        "type": "object"
                    },
                    "movement_type_detail": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Entrada"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "title": "Pagination Metadata",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "limit": {
                        "type": "integer",
                        "example": 25
                    },
                    "page_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "total_count": {
                        "type": "integer",
                        "example": 250
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}